requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
PyYAML
evidently
dill
matplotlib
kaggle
loguru
xgboost
catboost
mlxtend 
pymongo
pymongo[srv]
kneed
beautifulsoup4
seaborn
nltk
python-dotenv
dvc 
dvc[gdrive]
-e .

